Reorganize batch readers
